+++ src/Products/urban/Extensions/fixblobs.py
@@ -11,7 +11,6 @@ from plone import api
 blob_fields = {
     'SubTemplate': 'odt_file',
     'DashboardPODTemplate': 'odt_file',
-    'StyleTemplate': 'odt_file',
     'File': 'file',
     'UrbanTemplate': 'odt_file'
 }